Files/backend/requirements.txt
fastapi
uvicorn
yfinance
numpy
pandas>=1.5.0
python-multipart
pydantic
python-dotenv
requests